Refactoring routes for IT Support page in its own route .py file
