+++ app/templates/edit_ticket.html
@@ -186,7 +186,7 @@
 <div class="modal fade" id="deleteModal" tabindex="-1" aria-labelledby="deleteModalLabel" aria-hidden="true">
   <div class="modal-dialog modal-dialog-centered">
         <div class="modal-content">
-          <form method="post" action="{{ url_for('routes.delete_ticket', id=ticket.id) }}">
+          <form method="post" action="{{ url_for('troubleshooting.delete_ticket', id=ticket.id) }}">
             <div class="modal-header">
               <h5 class="modal-title" id="deleteModalLabel">Confirm Delete</h5>
               <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
